+++ packages/gallery/buildScripts/generateJSONSchemaFromTypes.js
@@ -1,16 +1,16 @@
-const path = require('path');
-const fs = require('fs');
+const path = require("path");
+const fs = require("fs");
 
-const TJS = require('typescript-json-schema');
+const TJS = require("typescript-json-schema")
 
 // optionally pass argument to schema generator
 const settings = {
-  required: true,
+    required: true,
 };
 
 // optionally pass ts compiler options
 const compilerOptions = {
-  strictNullChecks: true,
+    strictNullChecks: true,
 };
 
 // optionally pass a base path
@@ -18,16 +18,12 @@ const compilerOptions = {
 
 const program = TJS.getProgramFromFiles(
   [path.join(__dirname, '../src/components/gallery/styles.d.ts')],
-  compilerOptions
+  compilerOptions,
   // basePath
 );
 
 // We can either get the schema for one file and one type...
-const schema = TJS.generateSchema(program, 'StyleParams', settings);
-const schemaJSON = JSON.stringify(schema, null, 4);
-const template = `export default ${schemaJSON}`;
-fs.writeFileSync(
-  path.join(__dirname, '../src/components/gallery/typeValidator/schema.js'),
-  template,
-  { encoding: 'utf-8' }
-);
+const schema = TJS.generateSchema(program, "StyleParams", settings);
+const schemaJSON = JSON.stringify(schema, null, 4)
+const template = `export default ${schemaJSON}`
+fs.writeFileSync(path.join(__dirname, '../src/components/gallery/typeValidator/schema.js'), template, {encoding: 'utf-8'})