+++ packages/gallery/src/components/gallery/proGallery/slideshowView.js
@@ -74,21 +74,16 @@ class SlideshowView extends GalleryComponent {
   isScrollStart() {
     const { slideAnimation } = this.props.styleParams;
 
-    if (slideAnimation !== GALLERY_CONSTS.slideAnimations.SCROLL) {
+    if (slideAnimation !== GALLERY_CONSTS.slideAnimations.SCROLL || !this.scrollElement) {
       return false;
     }
-
-    if (this.container) {
-      return this.getScrollLeft() <= 1;
-    } else {
-      return false;
-    }
+    return this.scrollPosition() <= 1;
   }
 
   isScrollEnd() {
     const { totalItemsCount, getVisibleItems, galleryStructure, dimensions } =
       this.props;
-    const { slideshowLoop, slideAnimation } = this.props.styleParams;
+    const { slideshowLoop, slideAnimation, imageMargin } = this.props.styleParams;
 
     if (
       slideshowLoop ||
@@ -97,25 +92,15 @@ class SlideshowView extends GalleryComponent {
     ) {
       return false;
     }
+    const galleryStructureWidth = galleryStructure.width;
+    const visibleItemsCount = getVisibleItems(
+      galleryStructure.galleryItems,
+      dimensions
+    ).length;
+    const allItemsLoaded = visibleItemsCount >= totalItemsCount;
+    const scrollElementWidth = galleryStructureWidth - imageMargin / 2;
 
-    if (this.container) {
-      /* 'scrollWidth' is not reliable before first scroll (= it is equal to 'clientWidth' size).
-        'scrollWidth' will get his real value just after scrolling.
-      */
-      const { scrollWidth, clientWidth } = this.container;
-      const scrollLeft = this.getScrollLeft();
-      const visibleItemsCount = getVisibleItems(
-        galleryStructure.galleryItems,
-        dimensions
-      ).length;
-      const allItemsLoaded = visibleItemsCount >= totalItemsCount;
-      const visibleLeft = scrollLeft + clientWidth;
-      const visibleScroll = scrollWidth - 1;
-
-      return allItemsLoaded && visibleLeft >= visibleScroll;
-    } else {
-      return false;
-    }
+    return allItemsLoaded && this.scrollPositionAtTheAndOfTheGallery() >= scrollElementWidth;
   }
 
   isFirstItemFullyVisible() {
@@ -695,7 +680,9 @@ class SlideshowView extends GalleryComponent {
             const itemStyle = {
               width: thumbnailSize,
               height: thumbnailSize,
-              margin: thumbnailSpacings,
+              marginLeft: thumbnailSpacings,
+              marginTop: thumbnailSpacings,
+              overflow: 'hidden',
               backgroundImage: `url(${thumbnailItem.createUrl(
                 GALLERY_CONSTS.urlSizes.THUMBNAIL,
                 GALLERY_CONSTS.urlTypes.HIGH_RES
@@ -730,8 +717,13 @@ class SlideshowView extends GalleryComponent {
                   <TextItem
                     {...this.props}
                     {...thumbnailItem.renderProps()}
+                    styleParams={{
+                      ...this.props.styleParams,
+                      cubeType: 'fill',
+                      cubeImages: true
+                    }}
                     actions={{}}
-                    imageDimensions={itemStyle}
+                    imageDimensions={{...itemStyle, marginTop: 0, marginLeft: 0}}
                     style={{
                       ...thumbnailItem.renderProps().style,
                       ...itemStyle,
@@ -747,19 +739,20 @@ class SlideshowView extends GalleryComponent {
   }
 
   getCenteredItemIdxByScroll() {
-    const scrollLeft = this.getScrollLeft();
     // console.log('[RTL SCROLL] setCurrentItemByScroll: ', scrollLeft);
     const items = this.props.galleryStructure.galleryItems;
 
     let centeredIdx;
 
-    const scrollPos = scrollLeft + this.props.dimensions.galleryWidth / 2;
+    // const scrollPos = this.props.styleParams.isRTL ?
+    // this.props.galleryStructure.width - scrollLeft - this.props.container.galleryWidth / 2 :
+    const scrollPositionAtTheMiddleOfTheGallery = this.scrollPositionAtTheMiddleOfTheGallery();
 
-    if (scrollPos === 0) {
+    if (scrollPositionAtTheMiddleOfTheGallery === 0) {
       centeredIdx = 0;
     } else {
       for (let item, i = 0; (item = items[i]); i++) {
-        if (item.offset.left > scrollPos) {
+        if (item.offset.left > scrollPositionAtTheMiddleOfTheGallery) {
           centeredIdx = i - 1;
           break;
         }
@@ -772,19 +765,18 @@ class SlideshowView extends GalleryComponent {
   }
 
   getCenteredGroupIdxByScroll() {
-    const scrollLeft = this.getScrollLeft();
     // console.log('[RTL SCROLL] setCurrentItemByScroll: ', scrollLeft);
     const groups = this.props.galleryStructure.groups;
 
     let centeredGroupIdx;
 
-    const scrollPos = scrollLeft + this.props.dimensions.galleryWidth / 2;
+    const scrollPositionAtTheMiddleOfTheGallery = this.scrollPositionAtTheMiddleOfTheGallery();
 
-    if (scrollPos === 0) {
+    if (scrollPositionAtTheMiddleOfTheGallery === 0) {
       centeredGroupIdx = 0;
     } else {
       for (let group, i = 0; (group = groups[i]); i++) {
-        if (group.left > scrollPos) {
+        if (group.left > scrollPositionAtTheMiddleOfTheGallery) {
           centeredGroupIdx = i - 1;
           break;
         }
@@ -808,7 +800,7 @@ class SlideshowView extends GalleryComponent {
     }
 
     const isScrolling =
-      (this.container && this.container.getAttribute('data-scrolling')) ===
+      (this.scrollElement && this.scrollElement.getAttribute('data-scrolling')) ===
       'true';
 
     if (isScrolling) {
@@ -1319,10 +1311,22 @@ class SlideshowView extends GalleryComponent {
     };
   }
 
-  getScrollLeft() {
-    return this.container
-      ? (this.props.styleParams.isRTL ? -1 : 1) * this.container.scrollLeft
+  getScrollPosition() {
+    return this.scrollElement
+      ? this.scrollPosition()
       : 0;
+  }
+
+  scrollPositionAtTheMiddleOfTheGallery(){
+    return this.getScrollPosition() + this.props.dimensions.galleryWidth / 2;
+  }
+
+  scrollPositionAtTheAndOfTheGallery(){
+    return this.getScrollPosition() + this.props.dimensions.galleryWidth;
+  }
+
+  scrollPosition() {
+    return (this.props.styleParams.isRTL ? -1 : 1) * this.scrollElement.scrollLeft;
   }
 
   //-----------------------------------------| REACT |--------------------------------------------//
@@ -1423,7 +1427,6 @@ class SlideshowView extends GalleryComponent {
   }
 
   removeArrowsIfNeeded() {
-    setTimeout(() => {
       const { isRTL } = this.props.styleParams;
       const { hideLeftArrow, hideRightArrow } = this.state;
 
@@ -1437,18 +1440,18 @@ class SlideshowView extends GalleryComponent {
 
       const nextHideLeft = (!isRTL && atStart) || (isRTL && atEnd);
       const nextHideRight = (isRTL && atStart) || (!isRTL && atEnd);
-      const isNew =
+      const shouldUpdateArrowsState =
         !!nextHideLeft !== !!hideLeftArrow ||
         !!nextHideRight !== !!hideRightArrow;
 
-      if (isNew) {
+      if (shouldUpdateArrowsState) {
         this.setState({
           hideLeftArrow: !!nextHideLeft,
           hideRightArrow: !!nextHideRight,
         });
       }
-    }, 100);
   }
+
   navigationOutHandler() {
     //TODO remove after full refactor release
     utils.setStateAndLog(this, 'Next Item', {
@@ -1472,11 +1475,11 @@ class SlideshowView extends GalleryComponent {
     );
     window.addEventListener('gallery_navigation_in', this.navigationInHandler);
 
-    this.container = window.document.querySelector(
+    this.scrollElement = window.document.querySelector(
       `#pro-gallery-${this.props.domId} #gallery-horizontal-scroll`
     );
-    if (this.container) {
-      this.container.addEventListener('scroll', this._setCurrentItemByScroll);
+    if (this.scrollElement) {
+      this.scrollElement.addEventListener('scroll', this._setCurrentItemByScroll);
     }
     if (this.state.currentIdx > 0) {
       this.props.actions.scrollToItem(this.state.currentIdx);
@@ -1497,8 +1500,8 @@ class SlideshowView extends GalleryComponent {
       this.navigationInHandler
     );
 
-    if (this.container) {
-      this.container.removeEventListener(
+    if (this.scrollElement) {
+      this.scrollElement.removeEventListener(
         'scroll',
         this._setCurrentItemByScroll
       );