+++ packages/lib/src/settings/options/index.js
@@ -88,6 +88,7 @@ import pauseAutoSlideshowOnHover from './pauseAutoSlideshowOnHover';
 import showVideoControls from './showVideoControls';
 import shouldIndexDirectShareLinkInSEO from './shouldIndexDirectShareLinkInSEO';
 import slideTransition from './slideTransition';
+import useMaxDimensions from './useMaxDimensions';
 
 export default {
   itemClick,
@@ -180,6 +181,7 @@ export default {
   pauseAutoSlideshowOnHover,
   shouldIndexDirectShareLinkInSEO,
   slideTransition,
+  useMaxDimensions,
 };
 
 // TODO = add the options: