+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [v3.1.38](https://pro-gallery-3-1-38.surge.sh) (07/07/2021)
+ 
+#### GALLERY
+ -  fix: use imageDimensions in video item and text item
+
+---
+## [v3.1.37](https://pro-gallery-3-1-37.surge.sh) (07/07/2021)
+ 
+#### GALLERY
+ -  fix: Navigation Arrows Missing When Navigating to Page with Slider Gallery
+
+#### OTHER
+ -  feat:added alt text 
+
+---
+## [v3.1.36](https://pro-gallery-3-1-36.surge.sh) (06/07/2021)
+ 
+#### OTHER
+ -  improve: refactor image dimensions to layouter and added useMaxDImensions styleParam
+
+---
 ## [v3.1.35](https://pro-gallery-3-1-35.surge.sh) (04/07/2021)
  
 #### GALLERY